oh god big commit. tons of basic functionality
main page now displays a different random quote every time it loads. a fab shows a snackbar containing a random quote everytime it is pressed, and also now gives haptic feedback. a settings menu has been implemented that can either lead to a page that contains all quotes that exist and has a fab that displays a snackbar without functionality. a setting page allows you to change the interval, ringtone, and turn on and off notifications for new quotes. notifications not yet implemented. As well, a stock android settings page is included for demonstration.